requirements.txt
torch==2.6.0
torchvision==0.21.0
paddlepaddle==2.6.2
paddleocr==2.9.1
opencv-python-headless
Pillow
matplotlib
numpy
